Restructure....
